Create Knab.yml
So I made the score file for Knab, so we really have a list of all the Dutch banks with a banking license. I found almost all of the info on this page and in the pdf downloads: https://www.aegon.nl/overaegon/verantwoordbeleggen. It's incomplete, but this is what I could find in the pdf's. Feel free to expand and improve.